refactor to decrease code complexity remove unordered_incremental_executor_map
